giro-loader/requirements.txt
numpy==1.18.4
pandas==1.0.4
psycopg2-binary==2.8.5
python-dateutil==2.7.5
pytz==2020.1
six==1.15.0
SQLAlchemy==1.2.14
